Consolidate HDA and HDCA type names in client
