+++ client/src/api/datasets.ts
@@ -1,12 +1,10 @@
 import type { FetchArgType } from "openapi-typescript-fetch";
 
-import { DatasetDetails } from "@/api";
+import { HDADetailed } from "@/api";
 import { components, fetcher } from "@/api/schema";
 import { withPrefix } from "@/utils/redirect";
 
 export const datasetsFetcher = fetcher.path("/api/datasets").method("get").create();
-
-export type HDASummary = components["schemas"]["HDASummary"];
 
 type GetDatasetsApiOptions = FetchArgType<typeof datasetsFetcher>;
 type GetDatasetsQuery = Pick<GetDatasetsApiOptions, "limit" | "offset">;
@@ -42,11 +40,11 @@ export const fetchDataset = fetcher.path("/api/datasets/{dataset_id}").method("g
 
 export const fetchDatasetStorage = fetcher.path("/api/datasets/{dataset_id}/storage").method("get").create();
 
-export async function fetchDatasetDetails(params: { id: string }): Promise<DatasetDetails> {
+export async function fetchDatasetDetails(params: { id: string }): Promise<HDADetailed> {
     const { data } = await fetchDataset({ dataset_id: params.id, view: "detailed" });
     // We know that the server will return a DatasetDetails object because of the view parameter
     // but the type system doesn't, so we have to cast it.
-    return data as unknown as DatasetDetails;
+    return data as unknown as HDADetailed;
 }
 
 const updateDataset = fetcher.path("/api/datasets/{dataset_id}").method("put").create();
