+++ client/src/components/Dataset/DatasetList.vue
@@ -3,7 +3,8 @@ import { BAlert, BTable } from "bootstrap-vue";
 import { storeToRefs } from "pinia";
 import { computed, onMounted, ref } from "vue";
 
-import { copyDataset, getDatasets, HDASummary } from "@/api/datasets";
+import { HDASummary } from "@/api";
+import { copyDataset, getDatasets } from "@/api/datasets";
 import { updateTags } from "@/api/tags";
 import { useHistoryStore } from "@/stores/historyStore";
