+++ client/src/components/History/Content/Dataset/DatasetActions.vue
@@ -14,7 +14,7 @@ import { BButton } from "bootstrap-vue";
 import { computed } from "vue";
 import { useRouter } from "vue-router/composables";
 
-import { type DatasetDetails } from "@/api";
+import { type HDADetailed } from "@/api";
 import { copy as sendToClipboard } from "@/utils/clipboard";
 import localize from "@/utils/localization";
 import { absPath, prependPath } from "@/utils/redirect";
@@ -26,7 +26,7 @@ import DatasetDownload from "@/components/History/Content/Dataset/DatasetDownloa
 library.add(faBug, faChartBar, faInfoCircle, faLink, faQuestion, faRedo, faSitemap);
 
 interface Props {
-    item: DatasetDetails;
+    item: HDADetailed;
     writable: boolean;
     showHighlight: boolean;
     itemUrls: ItemUrls;
