+++ client/src/stores/datasetStore.ts
@@ -1,14 +1,14 @@
 import { defineStore } from "pinia";
 import { computed, set } from "vue";
 
-import type { DatasetDetails, DatasetEntry, HistoryContentItemBase } from "@/api";
+import type { DatasetEntry, HDADetailed, HistoryContentItemBase } from "@/api";
 import { fetchDataset } from "@/api/datasets";
 import { ApiResponse } from "@/api/schema";
 import { useKeyedCache } from "@/composables/keyedCache";
 
-async function fetchDatasetDetails(params: { id: string }): Promise<ApiResponse<DatasetDetails>> {
+async function fetchDatasetDetails(params: { id: string }): Promise<ApiResponse<HDADetailed>> {
     const response = await fetchDataset({ dataset_id: params.id, view: "detailed" });
-    return response as unknown as ApiResponse<DatasetDetails>;
+    return response as unknown as ApiResponse<HDADetailed>;
 }
 
 export const useDatasetStore = defineStore("datasetStore", () => {
